Phase1 (Data instertion to database)/requirements.txt
pyodbc==4.0.39
python-dotenv==1.0.0
streamlit==1.32.0
pandas==2.0.3
openpyxl==3.1.2
xlrd==2.0.1
plotly==5.18.0
matplotlib==3.8.0
